requirements.txt
chromadb
ollama
PyMuPDF  # for PDF processing
python-docx
requests
tqdm
fastapi
uvicorn
streamlit
